# perf: remove dynamic string creation for metric names