--- a/route/route.go
+++ b/route/route.go
@@ -98,6 +98,7 @@ type Router struct {
 
 	environmentCache *environmentCache
 	hsrv             *healthserver.Server
+	metricsMap       map[string]string
 }
 
 type BatchResponse struct {
@@ -168,7 +169,9 @@ func (r *Router) LnS(incomingOrPeer string) {
 
 	for _, metric := range routerMetrics {
 		if strings.HasPrefix(metric.Name, "_") {
-			metric.Name = r.incomingOrPeer + metric.Name
+			fullname := r.incomingOrPeer + metric.Name
+			r.metricsMap[metric.Name] = fullname
+			metric.Name = fullname
 		}
 		r.Metrics.Register(metric)
 	}
@@ -384,7 +387,7 @@ func (r *Router) marshalToFormat(w http.ResponseWriter, obj interface{}, format 
 
 // event is handler for /1/event/
 func (r *Router) event(w http.ResponseWriter, req *http.Request) {
-	r.Metrics.Increment(r.incomingOrPeer + "_router_event")
+	r.Metrics.Increment(r.metricsMap["_router_event"])
 
 	ctx := req.Context()
 	bodyBuffer, err := r.readAndCloseMaybeCompressedBody(req)
@@ -452,7 +455,7 @@ func (r *Router) requestToEvent(ctx context.Context, req *http.Request, reqBod [
 }
 
 func (r *Router) batch(w http.ResponseWriter, req *http.Request) {
-	r.Metrics.Increment(r.incomingOrPeer + "_router_batch")
+	r.Metrics.Increment(r.metricsMap["_router_batch"])
 
 	ctx := req.Context()
 	reqID := ctx.Value(types.RequestIDContextKey{})
@@ -539,7 +542,7 @@ func (router *Router) processOTLPRequest(
 	var requestID types.RequestIDContextKey
 	apiHost := router.Config.GetHoneycombAPI()
 
-	router.Metrics.Increment(router.incomingOrPeer + "_router_otlp")
+	router.Metrics.Increment(router.metricsMap["_router_otlp"])
 
 	// get environment name - will be empty for legacy keys
 	environment, err := router.getEnvironmentName(apiKey)
@@ -596,7 +599,7 @@ func (r *Router) processEvent(ev *types.Event, reqID interface{}) error {
 
 	if ev.Data.MetaTraceID == "" {
 		// not part of a trace. send along upstream
-		r.Metrics.Increment(r.incomingOrPeer + "_router_nonspan")
+		r.Metrics.Increment(r.metricsMap["_router_nonspan"])
 		debugLog.WithString("api_host", ev.APIHost).
 			WithString("dataset", ev.Dataset).
 			Logf("sending non-trace event from batch")
@@ -647,7 +650,7 @@ func (r *Router) processEvent(ev *types.Event, reqID interface{}) error {
 		// Figure out if we should handle this span locally or pass on to a peer
 		targetShard := r.Sharder.WhichShard(ev.Data.MetaTraceID)
 		if !targetShard.Equals(r.Sharder.MyShard()) {
-			r.Metrics.Increment(r.incomingOrPeer + "_router_peer")
+			r.Metrics.Increment(r.metricsMap["_router_peer"])
 			debugLog.
 				WithString("peer", targetShard.GetAddress()).
 				WithField("isprobe", isProbe).
@@ -677,12 +680,12 @@ func (r *Router) processEvent(ev *types.Event, reqID interface{}) error {
 		err = r.Collector.AddSpanFromPeer(span)
 	}
 	if err != nil {
-		r.Metrics.Increment(r.incomingOrPeer + "_router_dropped")
+		r.Metrics.Increment(r.metricsMap["_router_dropped"])
 		debugLog.Logf("Dropping span from batch, channel full")
 		return err
 	}
 
-	r.Metrics.Increment(r.incomingOrPeer + "_router_span")
+	r.Metrics.Increment(r.metricsMap["_router_span"])
 
 	debugLog.WithField("source", r.incomingOrPeer).Logf("Accepting span from batch for collection into a trace")
 	return nil