--- a/route/route.go
+++ b/route/route.go
@@ -98,7 +98,7 @@ type Router struct {
 
 	environmentCache *environmentCache
 	hsrv             *healthserver.Server
-	metricsMap       map[string]string
+	metricsMap       metrics.ComputedMetricNames
 }
 
 type BatchResponse struct {
@@ -167,14 +167,7 @@ func (r *Router) LnS(incomingOrPeer string) {
 		return
 	}
 
-	for _, metric := range routerMetrics {
-		if strings.HasPrefix(metric.Name, "_") {
-			fullname := r.incomingOrPeer + metric.Name
-			r.metricsMap[metric.Name] = fullname
-			metric.Name = fullname
-		}
-		r.Metrics.Register(metric)
-	}
+	r.metricsMap = metrics.NewComputedMetricNames(r.incomingOrPeer, routerMetrics)
 
 	muxxer := mux.NewRouter()
 
@@ -387,7 +380,7 @@ func (r *Router) marshalToFormat(w http.ResponseWriter, obj interface{}, format 
 
 // event is handler for /1/event/
 func (r *Router) event(w http.ResponseWriter, req *http.Request) {
-	r.Metrics.Increment(r.metricsMap["_router_event"])
+	r.Metrics.Increment(r.metricsMap.Get("_router_event"))
 
 	ctx := req.Context()
 	bodyBuffer, err := r.readAndCloseMaybeCompressedBody(req)
@@ -455,7 +448,7 @@ func (r *Router) requestToEvent(ctx context.Context, req *http.Request, reqBod [
 }
 
 func (r *Router) batch(w http.ResponseWriter, req *http.Request) {
-	r.Metrics.Increment(r.metricsMap["_router_batch"])
+	r.Metrics.Increment(r.metricsMap.Get("_router_batch"))
 
 	ctx := req.Context()
 	reqID := ctx.Value(types.RequestIDContextKey{})
@@ -475,7 +468,7 @@ func (r *Router) batch(w http.ResponseWriter, req *http.Request) {
 		r.handlerReturnWithError(w, ErrJSONFailed, err)
 		return
 	}
-	r.Metrics.Count(r.incomingOrPeer+"_router_batch_events", int64(len(batchedEvents.events)))
+	r.Metrics.Count(r.metricsMap.Get("_router_batch_events"), int64(len(batchedEvents.events)))
 
 	dataset, err := getDatasetFromRequest(req)
 	if err != nil {
@@ -542,7 +535,7 @@ func (router *Router) processOTLPRequest(
 	var requestID types.RequestIDContextKey
 	apiHost := router.Config.GetHoneycombAPI()
 
-	router.Metrics.Increment(router.metricsMap["_router_otlp"])
+	router.Metrics.Increment(router.metricsMap.Get("_router_otlp"))
 
 	// get environment name - will be empty for legacy keys
 	environment, err := router.getEnvironmentName(apiKey)
@@ -570,7 +563,7 @@ func (router *Router) processOTLPRequest(
 			}
 		}
 	}
-	router.Metrics.Count(router.incomingOrPeer+"_router_otlp_events", int64(totalEvents))
+	router.Metrics.Count(router.metricsMap.Get("_router_otlp_events"), int64(totalEvents))
 
 	return nil
 }
@@ -584,7 +577,7 @@ func (r *Router) processEvent(ev *types.Event, reqID interface{}) error {
 
 	// record the event bytes size
 	// we do this early so can include all event types (span, event, log, etc)
-	r.Metrics.Histogram(r.incomingOrPeer+"_router_event_bytes", float64(ev.GetDataSize()))
+	r.Metrics.Histogram(r.metricsMap.Get("_router_event_bytes"), float64(ev.GetDataSize()))
 
 	// An error here is effectively a parsing error, so return it up the stack.
 	if err := ev.Data.ExtractMetadata(); err != nil {
@@ -599,7 +592,7 @@ func (r *Router) processEvent(ev *types.Event, reqID interface{}) error {
 
 	if ev.Data.MetaTraceID == "" {
 		// not part of a trace. send along upstream
-		r.Metrics.Increment(r.metricsMap["_router_nonspan"])
+		r.Metrics.Increment(r.metricsMap.Get("_router_nonspan"))
 		debugLog.WithString("api_host", ev.APIHost).
 			WithString("dataset", ev.Dataset).
 			Logf("sending non-trace event from batch")
@@ -650,7 +643,7 @@ func (r *Router) processEvent(ev *types.Event, reqID interface{}) error {
 		// Figure out if we should handle this span locally or pass on to a peer
 		targetShard := r.Sharder.WhichShard(ev.Data.MetaTraceID)
 		if !targetShard.Equals(r.Sharder.MyShard()) {
-			r.Metrics.Increment(r.metricsMap["_router_peer"])
+			r.Metrics.Increment(r.metricsMap.Get("_router_peer"))
 			debugLog.
 				WithString("peer", targetShard.GetAddress()).
 				WithField("isprobe", isProbe).
@@ -680,12 +673,12 @@ func (r *Router) processEvent(ev *types.Event, reqID interface{}) error {
 		err = r.Collector.AddSpanFromPeer(span)
 	}
 	if err != nil {
-		r.Metrics.Increment(r.metricsMap["_router_dropped"])
+		r.Metrics.Increment(r.metricsMap.Get("_router_dropped"))
 		debugLog.Logf("Dropping span from batch, channel full")
 		return err
 	}
 
-	r.Metrics.Increment(r.metricsMap["_router_span"])
+	r.Metrics.Increment(r.metricsMap.Get("_router_span"))
 
 	debugLog.WithField("source", r.incomingOrPeer).Logf("Accepting span from batch for collection into a trace")
 	return nil