## Which problem is this PR solving?

When a string is dynamically created, it's allocated on the heap. This creates pressure on the GC which is biggest performance issue Refinery is currently battling with

## Short description of the changes

- Only compute necessary metric names once during startup
- store the computed result as struct fields for quicker access

## Benchmarks
```
**OLD**
goos: darwin
goarch: arm64
pkg: github.com/honeycombio/refinery/route
cpu: Apple M2 Max
BenchmarkRouterBatch-12    	5418073	     8647 ns/op	  17322 B/op	     41 allocs/op
PASS
ok  	github.com/honeycombio/refinery/route	47.190s

**NEW**
goos: darwin
goarch: arm64
pkg: github.com/honeycombio/refinery/route
cpu: Apple M2 Max
BenchmarkRouterBatch-12    	5644776	     8416 ns/op	  17092 B/op	     33 allocs/op
PASS
ok  	github.com/honeycombio/refinery/route	47.861s
```