@VinozzZ 2025-07-07
> Two things:
> 
> * Given this change, I think it makes sense to go further and change incomingOrPeer to either a bool (isIncoming) or an enum with two constants. There's little sense in carrying a string around when we don't use it for anything any more other than logging.
> * The change to the sampler constants is a good start, but their GetMetrics functions all do multiple string concatenations every call. It's probably worth an equivalent PR there to fix that -- and it might be worth changing it. Maybe we don't need to pass in the prefix to the samplers at all? At the very least, it could be passed in once rather than on every call.

I changed the `incomingOrPeer` into an enum with two constants.
I'm going to do a release for dynsampler-go and then update Refinery to the latest version in another PR